refactor: schema, resolver, dto 파일분리
