+++ server/src/dto/get-product-detail-dto.js
