+++ server/src/dto/product-search-dto.js
