+++ server/src/route.js
@@ -3,7 +3,7 @@ const router = express.Router()
 const { getErrorCode } = require('./errors')
 
 const { graphqlHTTP } = require('express-graphql')
-const schema = require('./schema')
+const schema = require('./schema/schema')
 
 router.get(
   '/graphql',
