+++ server/src/schema/mutation.js
@@ -1,64 +1,20 @@
 const {
-  GraphQLSchema,
   GraphQLObjectType,
   GraphQLID,
   GraphQLInt,
-  GraphQLString,
-  GraphQLList,
   GraphQLNonNull,
 } = require('graphql')
 
 const {
-  ProductType,
-  CartProductType,
-  ProdcutDetailImgType,
   changeStatusMessageType,
-} = require('./type')
+} = require('../type')
+
+const { likeProductResolver, dislikeProductResolver } = require('../resolver/like-resolver')
 const {
-  likeProductResolver,
-  dislikeProductResolver,
-  productListByCategoryResolver,
   addProductToCartResolver,
   modifyProductQuantityResolver,
   deleteProductFromCartResolver,
-  productListInCartResolver,
-  productDetalImgResolver
-} = require('./resolver')
-
-const RootQueryType = new GraphQLObjectType({
-  name: 'Query',
-  description: 'Root Query',
-  fields: () => ({
-    productListByCategory: {
-      type: new GraphQLList(ProductType),
-      description: 'List of products',
-      args: {
-        userId: { type: GraphQLNonNull(GraphQLID) },
-        category: { type: GraphQLString },
-        offset: { type: GraphQLInt },
-        limit: { type: GraphQLInt },
-        sorter: { type: GraphQLInt },
-      },
-      resolve: productListByCategoryResolver,
-    },
-    productListInCart: {
-      type: new GraphQLList(CartProductType),
-      description: '장바구니 상품 리스트',
-      args: {
-        userId: { type: GraphQLNonNull(GraphQLID) },
-      },
-      resolve: productListInCartResolver,
-    },
-    productDetailImg: {
-      type: new GraphQLList(ProdcutDetailImgType),
-      description: '상품 상세 이미지 src',
-      args: {
-        coupangProductId: { type: GraphQLNonNull(GraphQLInt) },
-      },
-      resolve: productDetalImgResolver,
-    },
-  }),
-})
+} = require('../resolver/cart-resolver')
 
 const RootMutationType = new GraphQLObjectType({
   name: 'Mutation',
@@ -113,9 +69,4 @@ const RootMutationType = new GraphQLObjectType({
   }),
 })
 
-const schema = new GraphQLSchema({
-  query: RootQueryType,
-  mutation: RootMutationType,
-})
-
-module.exports = schema
+module.exports = { RootMutationType }
